refactoring: unify transfer fragment for single and batch transfers
